issue-1122 fixed unit tests to work with FHIRRequestContext
