+++ fhir-persistence-jdbc/src/test/java/com/ibm/fhir/persistence/jdbc/test/spec/R4JDBCExamplesProcessor.java
@@ -6,23 +6,14 @@
 
 package com.ibm.fhir.persistence.jdbc.test.spec;
 
-import java.io.ByteArrayInputStream;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Properties;
 import java.util.function.Supplier;
 
-import javax.json.Json;
-import javax.json.JsonObject;
-import javax.json.JsonReader;
-import javax.json.JsonReaderFactory;
-
-import com.ibm.fhir.config.DefaultFHIRConfigProvider;
-import com.ibm.fhir.config.FHIRConfiguration;
+import com.ibm.fhir.config.FHIRConfigProvider;
 import com.ibm.fhir.config.FHIRRequestContext;
-import com.ibm.fhir.config.PropertyGroup;
 import com.ibm.fhir.database.utils.api.IConnectionProvider;
 import com.ibm.fhir.database.utils.api.ITransaction;
 import com.ibm.fhir.database.utils.api.ITransactionProvider;
@@ -41,7 +32,6 @@ import com.ibm.fhir.persistence.util.ResourceFingerprintVisitor;
  *
  */
 public class R4JDBCExamplesProcessor implements IExampleProcessor {
-    private static final JsonReaderFactory JSON_READER_FACTORY = Json.createReaderFactory(null);
 
     // the list of operations we apply to reach resource
     private final List<ITestResourceOperation> operations = new ArrayList<>();
@@ -60,6 +50,9 @@ public class R4JDBCExamplesProcessor implements IExampleProcessor {
     private String tenantName;
     private String tenantKey;
     private ITransactionProvider transactionProvider;
+    
+    // Adapter used by the persistence layer to access certain fhir-server-config properties
+    private final FHIRConfigProvider configProvider;
 
     /**
      * Public constructor. Uses a defaultlist of operations
@@ -67,10 +60,12 @@ public class R4JDBCExamplesProcessor implements IExampleProcessor {
      */
     public R4JDBCExamplesProcessor(FHIRPersistence persistence,
             Supplier<FHIRPersistenceContext> persistenceContextSupplier,
-            Supplier<FHIRPersistenceContext> historyContextSupplier) {
+            Supplier<FHIRPersistenceContext> historyContextSupplier,
+            FHIRConfigProvider configProvider) {
         this.persistence = persistence;
         this.persistenceContextSupplier = persistenceContextSupplier;
         this.historyContextSupplier = historyContextSupplier;
+        this.configProvider = configProvider;
 
         // The sequence of operations we want to apply to each resource
         operations.add(new CreateOperation());
@@ -83,6 +78,9 @@ public class R4JDBCExamplesProcessor implements IExampleProcessor {
         operations.add(new DeleteOperation());
         operations.add(new DeleteOperation());
         operations.add(new HistoryOperation(4)); // create+update+update+delete = 4 versions
+        
+        // Use a custom configuration provider so that we can support passing the tenant key
+        // without having to create a fhir-server-configuration.json file
     }
 
     /**
@@ -93,10 +91,11 @@ public class R4JDBCExamplesProcessor implements IExampleProcessor {
      * @param operations
      */
     public R4JDBCExamplesProcessor(FHIRPersistence persistence, Supplier<FHIRPersistenceContext> persistenceContextSupplier,
-            Supplier<FHIRPersistenceContext> historyContextSupplier, Collection<ITestResourceOperation> operations) {
+            Supplier<FHIRPersistenceContext> historyContextSupplier, Collection<ITestResourceOperation> operations, FHIRConfigProvider configProvider) {
         this.persistence = persistence;
         this.persistenceContextSupplier = persistenceContextSupplier;
         this.historyContextSupplier = historyContextSupplier;
+        this.configProvider = configProvider;
 
         // The sequence of operations we want to apply to each resource
         this.operations.addAll(operations);
@@ -110,7 +109,7 @@ public class R4JDBCExamplesProcessor implements IExampleProcessor {
      * @param operations
      */
     public R4JDBCExamplesProcessor(Collection<ITestResourceOperation> operations,
-            Properties configProps, IConnectionProvider cp, String tenantName, String tenantKey, ITransactionProvider transactionProvider) {
+            Properties configProps, IConnectionProvider cp, String tenantName, String tenantKey, ITransactionProvider transactionProvider, FHIRConfigProvider configProvider) {
         this.persistence = null;
         this.persistenceContextSupplier = null;
         this.historyContextSupplier = null;
@@ -119,42 +118,14 @@ public class R4JDBCExamplesProcessor implements IExampleProcessor {
         this.tenantName = tenantName;
         this.tenantKey = tenantKey;
         this.transactionProvider = transactionProvider;
+        this.configProvider = configProvider;
 
         // The sequence of operations we want to apply to each resource
         this.operations.addAll(operations);
     }
 
-    /**
-     * Configure the property group to inject the tenantKey, which is the only attribute
-     * required for this scenario
-     * @param configProvider
-     * @throws Exception
-     */
-    protected void configure(TestFHIRConfigProvider configProvider) throws Exception {
-
-        final String dsPropertyName = FHIRConfiguration.PROPERTY_DATASOURCES + "/default";
-
-        // The bare necessities we need to provide to the persistence layer in this case
-        final String jsonString = " {" + 
-                "    \"tenantKey\": \"" + this.tenantKey + "\"," + 
-                "    \"type\": \"db2\"," + 
-                "    \"multitenant\": true" + 
-                "}";
-        
-        try (JsonReader reader = JSON_READER_FACTORY.createReader(new ByteArrayInputStream(jsonString.getBytes(StandardCharsets.UTF_8)))) {
-            JsonObject jsonObj = reader.readObject();
-            PropertyGroup pg = new PropertyGroup(jsonObj);
-            configProvider.addPropertyGroup(dsPropertyName, pg);
-        }
-    }
-
     @Override
     public void process(String jsonFile, Resource resource) throws Exception {
-
-        // Use a custom configuration provider so that we can support passing the tenant key
-        // without having to create a fhir-server-configuration.json file
-        TestFHIRConfigProvider configProvider = new TestFHIRConfigProvider(new DefaultFHIRConfigProvider());
-        configure(configProvider);
 
         // Initialize the test context. As we run through the sequence of operations, each
         // one will update the context which will then be used by the next operation
